+++ app/javascript/packs/components/ChannelLine.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+// チャンネル表示
+const ChannelLine = props => (
+  <li key={props.key}>{props.textVal}</li>
+)
+
+export default ChannelLine
